requirements.txt
streamlit
snowflake-snowpark-python
supabase